+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2024-03-25
+
+### Added
+
+- MED-BBK dataset
+- Dataset statistics
+- Support for multi-source datasets
+- Multi-source case for MovieGraphBenchmark
+
+### Changed
+
+- entity links are now handled via eche's (Prefixed)ClusterHelper
+- Very large and very small datasets only allow dask/pandas backend respectively
+
+### Fixed
+
+- dask/pandas backend typing
+
 ## [0.2.1] - 2023-08-09
 
 ### Fixed